+++ wasmquery/handler.go
@@ -1,6 +1,7 @@
 package wasmquery
 
 import (
+	"context"
 	"fmt"
 	"time"
 
@@ -32,8 +33,7 @@ func newHandler(config *MethodConfig, enginePool *Engine, protoCodec Codec, logg
 	}, nil
 }
 
-func (h *Handler) handle(_ interface{}, stream grpc.ServerStream) error {
-	ctx := stream.Context()
+func (h *Handler) handle(_ interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 
 	logger := logging.Logger(ctx, h.logger)
 
@@ -44,8 +44,8 @@ func (h *Handler) handle(_ interface{}, stream grpc.ServerStream) error {
 
 	logger.Debug("handling wasm query call")
 	m := h.protoCodec.NewMessage()
-	if err := stream.RecvMsg(m); err != nil {
-		return err
+	if err := dec(m); err != nil {
+		return nil, err
 	}
 
 	vmInstance := h.enginePool.borrowVM(ctx)
@@ -61,17 +61,17 @@ func (h *Handler) handle(_ interface{}, stream grpc.ServerStream) error {
 
 	res, wasmErr, err := vmInstance.execute(request, h.exportName, m.Bytes())
 	if err != nil {
-		return fmt.Errorf("executing func %q: %w", h.exportName, err)
+		if vmInstance.panic != nil {
+			return nil, status.Error(codes.Internal, vmInstance.panic.Error())
+		}
+		return nil, status.Error(codes.Internal, fmt.Sprintf("unknown error executnig %q: %s", h.exportName, err))
 	}
-	if err, ok := wasmErr.(string); ok && err != "" {
-		return status.Error(codes.Internal, err)
+	if werr, ok := wasmErr.(string); ok && werr != "" {
+		return nil, status.Error(codes.Internal, werr)
 	}
 
 	out := h.protoCodec.NewMessage()
 	out.Set(res[0].([]byte))
 
-	if err := stream.SendMsg(out); err != nil {
-		return fmt.Errorf("send msg: %w", err)
-	}
-	return nil
+	return out, nil
 }